Fix for minor numerical discrepancy against blessed Bite output
On some systems, `make check` resulted in "912 deviations" for the Bite behavior, and none for others. The calculation that caused this discrepancy involved computing the square root of a very small number near zero. Perhaps due to differences in math libraries, this small number was on some systems computed as being slightly negative (though geometrically it should be strictly non-negative), and the sqrt would return nan. This would cause the `calcphi` and `musclelengthI2` functions to assume the odontophore was no longer in contact with I2, even when it was.

The blessed Bite output file was created on a system that computed with this error, so a new blessed Bite output file is included in this commit.